+++ src/test_cpverifier/library-checker/system_of_linear_equations.ss.test.cpp
@@ -11,7 +11,7 @@ using mat = tifa_libs::math::matrix<mint>;
 int main() {
   std::ios::sync_with_stdio(false);
   std::cin.tie(nullptr);
-  u64 n, m;
+  u32 n, m;
   std::cin >> n >> m;
   mat a(n, m), b(n, 1);
   std::cin >> a >> b;
